The version is now 8.2.0103. 1.Improve the handling of connection error messages. 2.Add an option to convert empty strings to NULL. 3.Handle domain type as the basetype. 4.Fake MSS in case of not only SQLExecute but also SQLParamData. 5.Prevent SQLColumns from displaying system columns. 6.Unload delayLoaded libraries on dll detach (not on disconnect time). 7.Fix a SQLGetDiagField crash bug. 8.Fix a insertion count bug. 9.Take win98 cases into account a little.

+++ socket.h
@@ -100,7 +100,7 @@ typedef void (WSAAPI *freeaddrinfo_func) (struct addrinfo *);
 typedef int (WSAAPI *getaddrinfo_func) (const char *, const char *,
 	const struct addrinfo *, struct addrinfo **); 
 typedef int (WSAAPI *getnameinfo_func) (const struct sockaddr *,
-	socklen_t, char *, DWORD, char *, DWORD, int);
+	socklen_t, char *, size_t, char *, size_t, int);
 
 #define SOCKET_ALREADY_CONNECTED		1
 #define SOCKET_HOST_NOT_FOUND			2
